[iOS] refactor: 풀스크린으로 촬영 화면 띄우기 & CaptureVC -> BaseVC
